Move commands to pkg/odo/cli.
