+++ pkg/odo/cli/cmdutils_test.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"fmt"
@@ -7,7 +7,7 @@ import (
 
 	appsv1 "github.com/openshift/api/apps/v1"
 	"github.com/redhat-developer/odo/pkg/occlient"
-	v1 "k8s.io/api/core/v1"
+	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/apimachinery/pkg/runtime"
